+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 <!-- For unreleased changes, see entries in .chloggen -->
 <!-- next version -->
 
+## [0.121.0] - 2025-03-18
+
+This Splunk OpenTelemetry Collector for Kubernetes release adopts the [Splunk OpenTelemetry Collector v0.121.0](https://github.com/signalfx/splunk-otel-collector/releases/tag/v0.121.0).
+
+### 🧰 Bug fixes 🧰
+
+- `other`: Fixed updating metrics' sourcetype with annotations ([#1375](https://github.com/signalfx/splunk-otel-collector-chart/pull/1375))
+
 ## [0.120.1] - 2025-03-07
 
 ### 💡 Enhancements 💡